pyproject.toml
[project]
name = "meta-ads-dashboard"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "chromadb>=1.1.0",
    "google-genai>=1.39.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.30",
    "langchain-openai>=0.3.34",
    "logfire>=4.11.0",
    "matplotlib>=3.10.6",
    "numpy>=2.2.6",
    "openai>=1.109.1",
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "plotly>=6.3.0",
    "pydantic-ai>=1.0.14",
    "python-dotenv>=1.1.1",
    "python-pptx>=0.6.23",
    "ics>=0.7.2",
    "requests>=2.32.5",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "seaborn>=0.13.2",
    "streamlit>=1.50.0",
    "pytest>=8.3.4",
]
